+++ src/app/components/file-browser/file-entry-charts.js
@@ -11,7 +11,6 @@ import Component from '@ember/component';
 import { get, computed, getProperties } from '@ember/object';
 import { reads } from '@ember/object/computed';
 import { inject as service } from '@ember/service';
-import { all as allFulfilled } from 'rsvp';
 import { promise } from 'ember-awesome-macros';
 import I18n from 'onedata-gui-common/mixins/components/i18n';
 import ColorGenerator from 'onedata-gui-common/utils/color-generator';
@@ -21,6 +20,7 @@ import { getTimeSeriesMetricNamesWithAggregator } from 'onedata-gui-common/utils
 import {
   dirSizeStatsTimeSeriesNameGenerators as timeSeriesNameGenerators,
 } from 'oneprovider-gui/models/file';
+import { hashSettled, hash as hashFulfilled, all as allFulfilled } from 'rsvp';
 
 const mixins = [
   I18n,
@@ -85,6 +85,13 @@ export default Component.extend(...mixins, {
   latestDirSizeStatsValuesUpdater: undefined,
 
   /**
+   * @type {ComputedProperty<string>}
+   */
+  currentProviderId: computed(function currentProviderId() {
+    return this.providerManager.getCurrentProviderId();
+  }),
+
+  /**
    * @type {ComputedProperty<PromiseObject<TimeSeriesCollectionSchema>>}
    */
   timeSeriesCollectionSchemaProxy: promise.object(computed(
@@ -94,22 +101,11 @@ export default Component.extend(...mixins, {
   )),
 
   /**
-   * @type {ComputedProperty<PromiseObject<number>>}
-   */
-  providersCountProxy: promise.object(computed(
-    'space',
-    async function providersCountProxy() {
-      return (await get(this.space, 'providerList')).hasMany('list').ids().length;
-    }
-  )),
-
-  /**
    * @type {ComputedProperty<PromiseObject>}
    */
   loadingProxy: promise.object(promise.all(
     'timeSeriesCollectionSchemaProxy',
     'latestDirSizeStatsValuesProxy',
-    'providersCountProxy'
   )),
 
   /**
@@ -158,6 +154,7 @@ export default Component.extend(...mixins, {
   fileCountChartSpec: computed(
     'seriesColorsConfig',
     'metricNamesForTimeSeries',
+    'currentProviderId',
     function fileCountChartSpec() {
       return {
         title: {
@@ -200,6 +197,7 @@ export default Component.extend(...mixins, {
                       data: {
                         externalSourceName: 'chartData',
                         externalSourceParameters: {
+                          collectionRef: this.currentProviderId,
                           timeSeriesNameGenerator: timeSeriesNameGenerators.dirCount,
                           timeSeriesName: timeSeriesNameGenerators.dirCount,
                           metricNames: this.metricNamesForTimeSeries
@@ -251,6 +249,7 @@ export default Component.extend(...mixins, {
                       data: {
                         externalSourceName: 'chartData',
                         externalSourceParameters: {
+                          collectionRef: this.currentProviderId,
                           timeSeriesNameGenerator: timeSeriesNameGenerators
                             .regFileAndLinkCount,
                           timeSeriesName: timeSeriesNameGenerators
@@ -295,6 +294,7 @@ export default Component.extend(...mixins, {
   sizeChartSpec: computed(
     'seriesColorsConfig',
     'metricNamesForTimeSeries',
+    'currentProviderId',
     function sizeChartSpec() {
       return {
         title: {
@@ -369,6 +369,7 @@ export default Component.extend(...mixins, {
                       data: {
                         externalSourceName: 'chartData',
                         externalSourceParameters: {
+                          collectionRef: this.currentProviderId,
                           timeSeriesNameGenerator: timeSeriesNameGenerators.totalSize,
                           timeSeriesName: timeSeriesNameGenerators.totalSize,
                           metricNames: this.metricNamesForTimeSeries
@@ -504,13 +505,24 @@ export default Component.extend(...mixins, {
   }),
 
   /**
+   * @type {ComputedProperty<DirCurrentSizeStatsResultForProvider|null>}
+   */
+  currentProviderLatestDirSizeStats: computed(
+    'latestDirSizeStatsValues',
+    'currentProviderId',
+    function currentProviderLatestDirSizeStats() {
+      return this.latestDirSizeStatsValues?.[this.currentProviderId] ?? null;
+    }
+  ),
+
+  /**
    * @type {ComputedProperty<string>}
    */
   stringifiedLatestElementsCount: computed(
-    'latestDirSizeStatsValues.{regFileAndLinkCount,dirCount}',
+    'currentProviderLatestDirSizeStats',
     function stringifiedLatestElementsCount() {
-      const fileCount = this.latestDirSizeStatsValues.regFileAndLinkCount ?? 0;
-      const dirCount = this.latestDirSizeStatsValues.dirCount ?? 0;
+      const fileCount = this.currentProviderLatestDirSizeStats?.regFileAndLinkCount ?? 0;
+      const dirCount = this.currentProviderLatestDirSizeStats?.dirCount ?? 0;
       const totalCount = fileCount + dirCount;
 
       const filesNounVer = fileCount === 1 ? 'singular' : 'plural';
@@ -529,20 +541,30 @@ export default Component.extend(...mixins, {
   ),
 
   /**
+   * @type {ComputedProperty<number>}
+   */
+  physicalSize: computed('latestDirSizeStatsValues', function physicalSize() {
+    return Object.values(this.latestDirSizeStatsValues || {}).reduce((acc, stats) => {
+      return acc + (stats.physicalSize ?? 0);
+    }, 0);
+  }),
+
+  /**
    * @type {ComputedProperty<string>}
    */
   physicalSizeOnProvidersDescription: computed(
-    'latestDirSizeStatsValues.physicalSizePerStorage',
-    'providersCountProxy.content',
+    'latestDirSizeStatsValues',
     function physicalSizeOnProvidersDescription() {
-      const providersCount = this.get('providersCountProxy.content');
-      if (!(providersCount > 1) || !this.latestDirSizeStatsValues) {
+      if (!this.latestDirSizeStatsValues) {
         return '';
       }
 
-      const providersWithStatsCount = Object.keys(
-        this.latestDirSizeStatsValues?.physicalSizePerStorage || {}
-      ).length;
+      const providersCount = Object.keys(this.latestDirSizeStatsValues).length;
+      const providersWithStatsCount = Object.keys(this.latestDirSizeStatsValues)
+        .filter((providerId) =>
+          this.latestDirSizeStatsValues[providerId].type === 'result'
+        )
+        .length;
 
       return this.t('currentSize.physicalSizeOnProvidersCount', {
         providersWithStatsCount,
@@ -555,14 +577,15 @@ export default Component.extend(...mixins, {
    * @type {ComputedProperty<string>}
    */
   physicalSizeExtraInfo: computed(
-    'latestDirSizeStatsValues.physicalSize',
+    'physicalSize',
     'physicalSizeOnProvidersDescription',
     function physicalSizeExtraInfo() {
       const extraInfo = [];
 
-      if (this.latestDirSizeStatsValues.physicalSize >= 1024) {
-        extraInfo.push(`(${this.latestDirSizeStatsValues.physicalSize} B)`);
+      if (this.physicalSize >= 1024) {
+        extraInfo.push(`(${this.physicalSize} B)`);
       }
+
       if (this.physicalSizeOnProvidersDescription) {
         extraInfo.push(this.physicalSizeOnProvidersDescription);
       }
@@ -618,55 +641,61 @@ export default Component.extend(...mixins, {
     metricNames,
   }) {
     const spaceId = get(this.space, 'entityId');
-    const timeSeriesNames = Object.keys(await this.getTimeSeriesCollectionLayout());
-    const dynamicSeries = await allFulfilled(
-      timeSeriesNames
-      .filter((timeSeriesName) => timeSeriesName.startsWith(timeSeriesNameGenerator))
-      .map(async (timeSeriesName) => {
-        const storageId = timeSeriesName.replace(timeSeriesNameGenerator, '');
-        let storageName = '';
-        let providerName = '';
-        let groupId;
-        try {
-          const storage = await this.storageManager.getStorageById(storageId, {
-            throughSpaceId: spaceId,
-            backgroundReload: false,
-          });
-          storageName = get(storage, 'name');
-          // Fetching provider record instead of just taking `entityId` from
-          // model relation to ensure, that provider is fetchable (and so
-          // calculated group_id will exist).
-          const provider = await get(storage, 'provider');
-          groupId = `provider_${get(provider, 'entityId')}`;
-          providerName = get(provider, 'name');
-        } catch (error) {
-          console.error(
-            `component:file-browser/file-entry-charts#fetchDynamicSeriesConfigs: cannot load storage with ID "${storageId}"`,
-            error
-          );
-          groupId = 'provider_unknown';
-        }
-        return {
-          id: storageId,
-          name: storageName || String(this.t('historicalSize.unknownStorage', {
-            id: storageId.slice(0, 6),
-          })),
-          groupId,
-          color: this.colorGenerator.generateColorForKey(storageId),
-          pointsSource: {
-            externalSourceName: 'chartData',
-            externalSourceParameters: {
-              timeSeriesNameGenerator,
-              timeSeriesName,
-              metricNames,
+    const dynamicSeriesPromises = [];
+    const collectionLayouts = await this.getTimeSeriesCollectionLayouts();
+    const providerIds = Object.keys(collectionLayouts);
+    const providers = await hashFulfilled(providerIds.reduce((acc, providerId) => {
+      acc[providerId] = this.providerManager.getProviderById(providerId, {
+        throughSpaceId: spaceId,
+      });
+      return acc;
+    }, {}));
+    providerIds.forEach((providerId) => {
+      const provider = providers[providerId];
+      const providerCollectionLayout = collectionLayouts[providerId];
+      const dynamicTimeSeriesNames = Object.keys(providerCollectionLayout)
+        .filter((timeSeriesName) => timeSeriesName.startsWith(timeSeriesNameGenerator));
+      dynamicTimeSeriesNames.forEach((timeSeriesName) => {
+        dynamicSeriesPromises.push((async () => {
+          const storageId = timeSeriesName.replace(timeSeriesNameGenerator, '');
+          let storageName = '';
+          const providerName = get(provider, 'name');
+          try {
+            const storage = await this.storageManager.getStorageById(storageId, {
+              throughSpaceId: spaceId,
+              backgroundReload: false,
+            });
+            storageName = get(storage, 'name');
+          } catch (error) {
+            console.error(
+              `component:file-browser/file-entry-charts#fetchDynamicSeriesConfigs: cannot load storage with ID "${storageId}"`,
+              error
+            );
+          }
+          return {
+            id: storageId,
+            name: storageName || String(this.t('historicalSize.unknownStorage', {
+              id: storageId.slice(0, 6),
+            })),
+            groupId: `provider_${providerId}`,
+            color: this.colorGenerator.generateColorForKey(storageId),
+            pointsSource: {
+              externalSourceName: 'chartData',
+              externalSourceParameters: {
+                collectionRef: providerId,
+                timeSeriesNameGenerator,
+                timeSeriesName,
+                metricNames,
+              },
             },
-          },
-          // Preparing sorting key in a way, that will move unknown providers at the end
-          // of series list and unknown storages at the end of provider series.
-          sortKey: `${providerName ? '\t' + providerName : '\n'}${storageName ? '\t' + storageName : '\n'}`,
-        };
-      })
-    );
+            // Preparing sorting key in a way, that will move unknown storages
+            // at the end of provider series.
+            sortKey: `${providerName}${storageName ? '\t' + storageName : '\n'}`,
+          };
+        })());
+      });
+    });
+    const dynamicSeries = await allFulfilled(dynamicSeriesPromises);
     return dynamicSeries.sortBy('sortKey');
   },
 
@@ -702,10 +731,24 @@ export default Component.extend(...mixins, {
   },
 
   /**
-   * @returns {Promise<TimeSeriesCollectionLayout>}
+   * @returns {Promise<Object<string, TimeSeriesCollectionLayout>>}
+   *   map providerId -> TS collection layout for that provider
    */
-  async getTimeSeriesCollectionLayout() {
-    return this.fileManager.getDirSizeStatsTimeSeriesCollectionLayout(this.fileId);
+  async getTimeSeriesCollectionLayouts() {
+    const providerList = await get(this.space, 'providerList');
+    const providers = await get(providerList, 'list');
+    const allProvidersLayouts = await hashSettled(providers.reduce((acc, provider) => {
+      const providerId = get(provider, 'entityId');
+      acc[providerId] = this.fileManager
+        .getDirSizeStatsTimeSeriesCollectionLayout(this.fileId, providerId);
+      return acc;
+    }, {}));
+    return Object.keys(allProvidersLayouts).reduce((acc, providerId) => {
+      if (allProvidersLayouts[providerId].state === 'fulfilled') {
+        acc[providerId] = allProvidersLayouts[providerId].value;
+      }
+      return acc;
+    }, {});
   },
 
   /**
@@ -737,6 +780,7 @@ export default Component.extend(...mixins, {
 
       const slice = await this.fileManager.getDirSizeStatsTimeSeriesCollectionSlice(
         this.fileId,
+        batchedQuery.collectionRef,
         queryParams
       );